requirements.txt
Django
SemanticScholar
python-dotenv
neomodel
neo4j
spacy
rdflib
django-unicorn